Refactor journal contexts
