+++ internal/journal/contexts/dummy.lua
@@ -1,0 +1,13 @@
+--@ module = true
+
+-- Dummy Context, no storage --
+
+DummyJournalContext = defclass(DummyJournalContext)
+
+function DummyJournalContext:save(text, cursor)
+end
+
+function DummyJournalContext:load()
+  return {text={''}, cursor={1}, show_tutorial=true}
+end
+
